polkit: fix CVE-2025-7519
Source: meta-openembedded
MR: 296370
Type: Integration
Disposition: Merged from meta-openembedded
ChangeID: 0d88144a27e2222366c8f3f2fcc50623f96afe7d
Description:

A flaw was found in polkit. When processing an XML policy with 32 or
more nested elements in depth, an out-of-bounds write can be triggered.
This issue can lead to a crash or other unexpected behavior, and
arbitrary code execution is not discarded. To exploit this flaw, a
high-privilege account is needed as it's required to place the
malicious policy file properly.

Reference:
https://nvd.nist.gov/vuln/detail/CVE-2025-7519

Upstream-patch:
https://github.com/polkit-org/polkit/commit/107d3801361b9f9084f78710178e683391f1d245

Signed-off-by: Praveen Kumar <praveen.kumar@windriver.com>
Signed-off-by: Anuj Mittal <anuj.mittal@intel.com>
Signed-off-by: Jeremy A. Puhlman <jpuhlman@mvista.com>