fontforge: patch CVE-2020-5395, CVE-2020-25690 and CVE-2020-5496
Source: meta-openembedded
MR: 302734
Type: Integration
Disposition: Merged from meta-openembedded
ChangeID: 2491ea2ffb7d1b8f5025b41b949b7d7bce02ab81
Description:

Details: https://nvd.nist.gov/vuln/detail/CVE-2020-5395
https://nvd.nist.gov/vuln/detail/CVE-2020-25690
https://nvd.nist.gov/vuln/detail/CVE-2020-5496

The same patch fixes all three.
The patch for CVE-2020-25690 is mentioned in the RedHat bug, which is
referenced in the nvd report.
The patch for CVE-2020-5395 is mentioned in the Github issue that
is referenced in the nvd report.
The patch for CVE-2020-5496 is mentioned in the comments of the issue
that is linked in the nvd report.

Signed-off-by: Gyorgy Sarvari <skandigraun@gmail.com>
Signed-off-by: Jeremy A. Puhlman <jpuhlman@mvista.com>