pcp: switch SRC_URI to git
Source: meta-openembedded
MR: 302734
Type: Integration
Disposition: Merged from meta-openembedded
ChangeID: 02422765c6a3c5a51c65dd8cf86b29543fe363cd
Description:

The original link stopped working.

I have compared the original tarball's content with this revision: the contents
are bit-identical to each other. The only difference is that the original
tarball came with an extra "debian/control" file which is not present in
the git repository, but it not using for compiling.

Signed-off-by: Gyorgy Sarvari <skandigraun@gmail.com>
Signed-off-by: Jeremy A. Puhlman <jpuhlman@mvista.com>